Writing some TODO-s. And fixing some error HTTP response codes.

+++ src/main/java/org/example/user_management_application/web/UserController.java
@@ -18,20 +18,25 @@ public class UserController {
         this.userService = userService;
     }
 
-    //TODO: (РАБОТИ) Вади всички Юзъри от базата.
+    //TODO: (РАБОТИ)
+    // Вади всички Юзъри от базата - връща код "Oк" (200).
+    // Ако листа е празен връща - код "No Content" (204).
+    // Листа е подреден 1-во по Фамилно име и 2-ро по Дата на раждане.
     @GetMapping("/api/users")
     public ResponseEntity<List<User>> getUsers() {
 
         List<User> allUser = userService.getAllUser();
 
         if (allUser.isEmpty()) {
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.noContent().build();
         }
-
         return ResponseEntity.ok(allUser);
+
     }
 
-    //TODO: (РАБОТИ) Вади конкретен Юзър по ID от базата.
+    //TODO: (РАБОТИ)
+    // Показва конкретен Юзър по ID от базата - връща код "Oк" (200).
+    // Ако юзъра не съществува връща - код "No Content" (204).
     @GetMapping("/api/user/{userId}")
     public ResponseEntity<Optional<User>> getUser(@PathVariable Long userId) {
 
@@ -40,12 +45,13 @@ public class UserController {
         if (userById.isEmpty()) {
             return ResponseEntity.noContent().build();
         }
-
         return ResponseEntity.ok(userById);
     }
 
     //TODO: (РАБОТИ)
-    //TODO: CREATE USER !!! Създава Юзър като подаваме неговото ИД, Първо име, Последно име, Дата на раждане(формат - 2007-12-03 година-месец-ден), Телефон и Имейл.
+    // Създава Юзър като подаваме неговото ИД, Първо име, Последно име, Дата на раждане(формат - 2007-12-03 година-месец-ден), Телефон и Имейл.
+    // Ако няма такъв обект в базата го създава и връща код "Created" (201).
+    // Ако има такъв обект връща код "IM Used" (226).
     @PostMapping("/api/create/{id}/{firstName}/{lastName}/{dateOfBirth}/{phoneNumber}/{email}")
     public ResponseEntity<User> createUser(@PathVariable Long id,
                                            @PathVariable String firstName,
@@ -65,6 +71,7 @@ public class UserController {
                 .setEmail(email);
 
         boolean existOrNot = userService.isUserExist(userBindingModel);
+
         if (!existOrNot) {
             User user = userService.createUser(userBindingModel);
             return new ResponseEntity<>(user, HttpStatus.CREATED);
@@ -73,18 +80,36 @@ public class UserController {
 
     }
 
-    //TODO: (РАБОТИ) Трие Юзър при подадено ID.
+    //TODO: (РАБОТИ)
+    // Трие Юзър при подадено ID.
+    // Ако Юзъра съществува в базата се изтрива и връща код "No Content" (204).
+    // Ако го няма в базата връща код "Not Found" (404).
     @DeleteMapping("/api/delete/{id}")
-    public void deleteUser(@PathVariable Long id) {
-        userService.deleteUser(id);
+    public ResponseEntity<User> deleteUser(@PathVariable Long id) {
+
+        boolean exist = userService.userExist(id);
+
+        if (exist) {
+            userService.deleteUser(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 
     }
 
     //TODO: (РАБОТИ) Ъпдейтва Юзър при подадено ID.
+    // Ако Юзъра съществува в базата връща код "Ок" (200).
+    // Ако го няма в базата връща код "Not Found" (404).
     @PutMapping("/api/update/{id}")
     public ResponseEntity<User> updateUser(@PathVariable Long id, @RequestBody User user) {
-        userService.updateUser(id, user);
-        return ResponseEntity.ok(user);
-    }
 
+        boolean exist = userService.userExist(id);
+
+        if (exist) {
+            userService.updateUser(id, user);
+            return ResponseEntity.ok(user);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+    }
 }
